pipeline/cloud/services/ocr/requirements.txt
google-cloud-vision
google-cloud-storage
